rename : 온보딩 알림 설정 - 명확한 이름으로 변경
