+++ app/src/main/java/com/dongyang/android/youdongknowme/ui/view/permission/OnboardingPermission.kt
@@ -1,4 +1,4 @@
-package com.dongyang.android.youdongknowme.ui.view.settingAlarm
+package com.dongyang.android.youdongknowme.ui.view.permission
 
 import android.content.Intent
 import android.os.Build
@@ -8,15 +8,15 @@ import android.text.style.ForegroundColorSpan
 import android.widget.TextView
 import androidx.annotation.RequiresApi
 import com.dongyang.android.youdongknowme.R
-import com.dongyang.android.youdongknowme.databinding.ActivitySettingAlarmBinding
+import com.dongyang.android.youdongknowme.databinding.ActivityOnboardingPermissionBinding
 import com.dongyang.android.youdongknowme.standard.base.BaseActivity
 import com.dongyang.android.youdongknowme.ui.view.main.MainActivity
 import com.dongyang.android.youdongknowme.ui.view.setting.SettingViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class SettingAlarmActivity : BaseActivity<ActivitySettingAlarmBinding, SettingViewModel>() {
+class OnboardingPermission : BaseActivity<ActivityOnboardingPermissionBinding, SettingViewModel>() {
 
-    override val layoutResourceId: Int = R.layout.activity_setting_alarm
+    override val layoutResourceId: Int = R.layout.activity_onboarding_permission
     override val viewModel: SettingViewModel by viewModel()
 
     override fun initStartView() {
@@ -28,7 +28,7 @@ class SettingAlarmActivity : BaseActivity<ActivitySettingAlarmBinding, SettingVi
     @RequiresApi(Build.VERSION_CODES.P)
     override fun initAfterBinding() {
         binding.btnAlarmComplete.setOnClickListener {
-            val intent = Intent(this@SettingAlarmActivity, MainActivity::class.java)
+            val intent = Intent(this@OnboardingPermission, MainActivity::class.java)
             startActivity(intent)
             finish()
         }
